+++ CoEpi/storage/DBCENKey.swift
@@ -1,7 +1,7 @@
 import RealmSwift
 
 final class DBCENKey : Object {
-    @objc dynamic var timestamp: Int = Int(Date().timeIntervalSince1970)
+    @objc dynamic var timestamp: Int64 = Int64(Date().timeIntervalSince1970)
     @objc dynamic var CENKey: String = ""
     var latitude: Double? = nil
     var longitude: Double? = nil
@@ -10,14 +10,14 @@ final class DBCENKey : Object {
         return "timestamp"
     }
     
-    required init( _ts: Int, _cenKey: String ) {
+    required init( _ts: Int64, _cenKey: String ) {
         self.timestamp = _ts
         self.CENKey = _cenKey
     }
     
     required init() {
         //fatalError("init() has not been implemented")
-        self.timestamp = Int(Date().timeIntervalSince1970)
+        self.timestamp = Int64(Date().timeIntervalSince1970)
         self.CENKey = ""
     }
 }
